Add additonal controls to oscilloscopes
Display of the second parameter is still not working, probably because
ether std::to_string is not happy with floats, or the text is too long
and the display code is breaking it. Either way, a custom tostring
may be required.

+++ oscilloscope.hpp
@@ -44,8 +44,8 @@ public:
 
     void Init(OscilloscopeParams& params);
 
-    unsigned int GetGain();
-    void SetGain(unsigned int gain);
+    float GetGain();
+    void SetGain(float gain);
 
     unsigned int GetScale();
     void SetScale(unsigned int xScale);
